+++ index.php
@@ -2,9 +2,9 @@
 /**
  * SyncMarks
  *
- * @version 1.6.7.1
+ * @version 1.6.8
  * @author Offerel
- * @copyright Copyright (c) 2021, Offerel
+ * @copyright Copyright (c) 2022, Offerel
  * @license GNU General Public License, version 3
  */
 session_start();
@@ -235,7 +235,7 @@ if(isset($_POST['action'])) {
 			switch($format) {
 				case "html":
 					e_log(8,"Exporting in HTML format for download");
-					die(html_export());
+					sendJSONResponse(html_export());
 					break;
 				case "json":
 					e_log(8,"Exporting in JSON format");
@@ -245,7 +245,7 @@ if(isset($_POST['action'])) {
 						file_put_contents($filename,$bookmarks,true);
 						e_log(8,"Export file is saved to $filename");
 					}
-					$bcount = count(json_decode($bookmarks));
+					$bcount = count(json_decode($bookmarks)) + 1;
 					e_log(8,"Send $bcount bookmarks to '$client'");
 					$ctime = (filter_var($_POST['sync'], FILTER_SANITIZE_STRING) === 'false') ? 0:$ctime;
 					updateClient($client, $ctype, $ctime, true);
@@ -302,18 +302,30 @@ if(isset($_POST['action'])) {
 			break;
 		case "addmark":
 			$bookmark = json_decode($_POST['data'], true);
-			e_log(8,"Try to add new bookmark '".$bookmark['title']."'");
 			$stime = round(microtime(true) * 1000);
 			$bookmark['added'] = $stime;
-			$bookmark['title'] = htmlspecialchars(mb_convert_encoding(htmlspecialchars_decode($bookmark['title'], ENT_QUOTES),"UTF-8"),ENT_QUOTES,'UTF-8', false);
+			$bookmark['title'] = ($bookmark['title'] === '') ? getSiteTitle($bookmark['url']):htmlspecialchars(mb_convert_encoding(htmlspecialchars_decode($bookmark['title'], ENT_QUOTES),"UTF-8"),ENT_QUOTES,'UTF-8', false);
+			e_log(8,"Try to add new bookmark '".$bookmark['title']."'");
 			e_log(9, print_r($bookmark, true));
 			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
 			if(array_key_exists('url',$bookmark)) $bookmark['url'] = validate_url($bookmark['url']);
 			if(strtolower(getClientType($_SERVER['HTTP_USER_AGENT'])) != "firefox") $bookmark = cfolderMatching($bookmark);
 			$stime = (!isset($_POST['sync'])) ? 0:$stime;
+			
 			if($bookmark['type'] == 'bookmark' && isset($bookmark['url'])) {
 				$response = addBookmark($bookmark);
-				updateClient($client, strtolower(getClientType($_SERVER['HTTP_USER_AGENT'])), $stime, true);
+
+				if($_POST['add'] === '2') {
+					if($response === 1) {
+						e_log(8,"Bookmark added");
+						sendJSONResponse(bmTree());
+					} else {
+						echo $res;
+						http_response_code(417);
+					}
+				} else {
+					updateClient($client, strtolower(getClientType($_SERVER['HTTP_USER_AGENT'])), $stime, true);
+				}
 			} else if($bookmark['type'] == 'folder') {
 				$response = addFolder($bookmark);
 				updateClient($client, strtolower(getClientType($_SERVER['HTTP_USER_AGENT'])), $stime, true);
@@ -355,10 +367,11 @@ if(isset($_POST['action'])) {
 			} else {
 				$query = "SELECT `bmID` FROM `bookmarks` WHERE `bmType` = 'folder' AND `bmTitle` = '".$bookmark['title']."' AND `userID` = ".$_SESSION['sud']['userID'].";";
 			}
+			$response = 1;
 			$bData = db_query($query);
 			if(count($bData) == 1) {
-				e_log(2, "Bookmark found, trying to remove it");
-				$response = delMark($bData[0]['bmID']);
+				e_log(8, "Bookmark found, trying to remove it");
+				$response = delMark(array($bData[0]['bmID']));
 			} else if (count($bData) > 1) {
 				$response = "No unique bookmark found, doing nothing";
 				e_log(2, $response);
@@ -376,35 +389,34 @@ if(isset($_POST['action'])) {
 			sendJSONResponse(db_query($query));
 			break;
 		case "arename":
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
+			$client = (isset($_POST['add']) && $_POST['add'] != 'null') ? filter_var($_POST['add'], FILTER_SANITIZE_STRING):filter_var($_POST['client'], FILTER_SANITIZE_STRING);
 			$name = filter_var($_POST['data'], FILTER_SANITIZE_STRING);
 			e_log(8,"Rename client $client to $name");
 			$query = "UPDATE `clients` SET `cname` = '".$name."' WHERE `uid` = ".$_SESSION['sud']['userID']." AND `cid` = '".$client."';";
 			$count = db_query($query);
 			$response = ($count > 0) ? bClientlist($_SESSION['sud']['userID']):false;
-			
 			sendJSONResponse($response);
 			break;
 		case "cfolder":
 			$ctime = round(microtime(true) * 1000);
-			$fname = filter_var($_POST['fname'], FILTER_SANITIZE_STRING);
-			$fbid = filter_var($_POST['fbid'], FILTER_SANITIZE_STRING);
-			die(cfolder($ctime,$fname,$fbid));
+			$fname = filter_var($_POST['data'], FILTER_SANITIZE_STRING);
+			$fbid = filter_var($_POST['add'], FILTER_SANITIZE_STRING);
+			sendJSONResponse(cfolder($ctime,$fname,$fbid));
 			break;
 		case "rmessage":
-			$message = isset($_POST['message']) ? filter_var($_POST['message'], FILTER_VALIDATE_INT):0;
-			$loop = filter_var($_POST['lp'], FILTER_SANITIZE_STRING) == 'aNoti' ? 1:0;
+			$message = isset($_POST['data']) ? filter_var($_POST['data'], FILTER_VALIDATE_INT):0;
+			$loop = filter_var($_POST['add'], FILTER_SANITIZE_STRING) == 'aNoti' ? 1:0;
 			if($message > 0) {
 				e_log(8,"Try to delete notification $message");
 				$query = "DELETE FROM `notifications` WHERE `userID` = ".$_SESSION['sud']['userID']." AND `id` = $message;";
 				$count = db_query($query);
 				($count === 1) ? e_log(8,"Notification successfully removed") : e_log(9,"Error, removing notification");
 			}
-			die(notiList($_SESSION['sud']['userID'], $loop));
+			sendJSONResponse(notiList($_SESSION['sud']['userID'], $loop));
 			break;
 		case "soption":
-			$option = filter_var($_POST['option'], FILTER_SANITIZE_STRING);
-			$value = filter_var(filter_var($_POST['value'], FILTER_SANITIZE_NUMBER_INT), FILTER_VALIDATE_INT);
+			$option = filter_var($_POST['data'], FILTER_SANITIZE_STRING);
+			$value = filter_var(filter_var($_POST['add'], FILTER_SANITIZE_NUMBER_INT), FILTER_VALIDATE_INT);
 			e_log(8,"Option received: ".$option.":".$value);
 			$oOptionsA = json_decode($_SESSION['sud']['uOptions'],true);
 			$oOptionsA[$option] = $value;
@@ -412,10 +424,10 @@ if(isset($_POST['action'])) {
 			header("Content-Type: application/json");
 			if(db_query($query) !== false) {
 				e_log(8,"Option saved");
-				die(json_encode(true));
+				sendJSONResponse(json_encode(true));
 			} else {
 				e_log(9,"Error, saving option");
-				die(json_encode(false));
+				sendJSONResponse(json_encode(false));
 			}
 			break;
 		case "tl":
@@ -450,30 +462,31 @@ if(isset($_POST['action'])) {
 			die(json_encode($tResponse));
 			break;
 		case "bmedt":
-			$title = filter_var($_POST['title'], FILTER_SANITIZE_STRING);
-			$id = filter_var($_POST['id'], FILTER_SANITIZE_STRING);
-			e_log(8,"Edit entry '$title'");
-			$url = (isset($_POST['url']) && strlen($_POST['url']) > 4) ? '\''.validate_url($_POST['url']).'\'' : 'NULL';
+			$bookmark = json_decode($_POST['data'], true);
+			$title = filter_var($bookmark['title'], FILTER_SANITIZE_STRING);
+			$id = filter_var($bookmark['id'], FILTER_SANITIZE_STRING);
+			$url = (isset($bookmark['url']) && strlen($bookmark['url']) > 4) ? '\''.validate_url($bookmark['url']).'\'' : 'NULL';
+			e_log(8, "Edit entry '$title'");
 			$query = "UPDATE `bookmarks` SET `bmTitle` = '$title', `bmURL` = $url, `bmAdded` = '".round(microtime(true) * 1000)."' WHERE `bmID` = '$id' AND `userID` = ".$_SESSION['sud']['userID'].";";
 			$count = db_query($query);
-			($count > 0) ? die(true) : die(false);
+			($count > 0) ? sendJSONResponse(true):sendJSONResponse(false);
 			break;
 		case "bmmv":
-			$id = filter_var($_POST['id'], FILTER_SANITIZE_STRING);
+			$id = filter_var($_POST['add'], FILTER_SANITIZE_STRING);
 			e_log(8,"Move bookmark $id");
-			$folder = filter_var($_POST['folder'], FILTER_SANITIZE_STRING);
-			$query = "SELECT MAX(bmIndex)+1 AS 'index' FROM `bookmarks` WHERE `bmParentID` = '$folder';";
+			$folder = filter_var($_POST['data'], FILTER_SANITIZE_STRING);
+			$query = "SELECT IFNULL(MAX(bmIndex), 0) + 1 AS 'index' FROM `bookmarks` WHERE `bmParentID` = '$folder';";
 			$folderData = db_query($query);
 			$query = "UPDATE `bookmarks` SET `bmIndex` = ".$folderData[0]['index'].", `bmParentID` = '$folder', `bmAdded` = '".round(microtime(true) * 1000)."' WHERE `bmID` = '$id' AND `userID` = ".$_SESSION['sud']['userID'].";";
 			$count = db_query($query);
-			($count > 0) ? die(true) : die(false);
+			($count > 0) ? sendJSONResponse(true):sendJSONResponse(false);
 			break;
 		case "adel":
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
+			$client = filter_var($_POST['data'], FILTER_SANITIZE_STRING);
 			e_log(8,"Delete client $client");
 			$query = "DELETE FROM `clients` WHERE `uid` = ".$_SESSION['sud']['userID']." AND `cid` = '$client';";
 			$count = db_query($query);
-			($count > 0) ? die(bClientlist($_SESSION['sud']['userID'])) : die(false);
+			($count > 0) ? sendJSONResponse(bClientlist($_SESSION['sud']['userID'])):sendJSONResponse(false);
 			break;
 		case "cmail":
 			e_log(8,"Change e-mail for ".$_SESSION['sud']['userName']);
@@ -496,10 +509,13 @@ if(isset($_POST['action'])) {
 			$del = false;
 			$headers = "From: SyncMarks <".CONFIG['sender'].">";
 			$url = $_SERVER['REQUEST_SCHEME']."://".$_SERVER['HTTP_HOST'].$_SERVER['SCRIPT_NAME'];
-			$variant = filter_var($_POST['type'], FILTER_VALIDATE_INT);
-			$password = (isset($_POST['p']) && $_POST['p'] != '') ? filter_var($_POST['p'], FILTER_SANITIZE_STRING):gpwd(16);
-			$userLevel = filter_var($_POST['userLevel'], FILTER_VALIDATE_INT);
-			$user = filter_var($_POST['nuser'], FILTER_SANITIZE_STRING);
+
+			$data = json_decode($_POST['data'], true);
+
+			$variant = filter_var($data['type'], FILTER_VALIDATE_INT);
+			$password = (isset($data['p']) && $data['p'] != '') ? filter_var($data['p'], FILTER_SANITIZE_STRING):gpwd(16);
+			$userLevel = filter_var($data['userLevel'], FILTER_VALIDATE_INT);
+			$user = filter_var($data['nuser'], FILTER_SANITIZE_STRING);
 			$mail = filter_var($user, FILTER_VALIDATE_EMAIL) ? $user:null;
 
 			switch($variant) {
@@ -511,7 +527,7 @@ if(isset($_POST['action'])) {
 					if($nuid > 0) {
 						if(filter_var($mail, FILTER_VALIDATE_EMAIL)) {
 							$response = $nuid;
-							$message = "Hello,\r\na new account with the following credentials is created and stored encrypted on for SyncMarks:\r\nUsername: $user\r\nPassword: $password\r\n\r\nYou can login at $url";
+							$message = "Hello,\r\na new account with the following credentials is created for SyncMarks:\r\nUsername: $user\r\nPassword: $password\r\n\r\nYou can login at $url";
 							if(!mail ($mail, "Account created",$message,$headers)) {
 								e_log(1,"Error sending data for created user account to user");
 								$response = "User created successful, E-Mail could not send";
@@ -527,12 +543,10 @@ if(isset($_POST['action'])) {
 					} else {
 						$response = "User creation failed";
 					}
-					header("Content-Type: application/json");
-					die(json_encode($response));
 					break;
 				case 2:
 					e_log(8,"Updating user $user");
-					$uID = filter_var($_POST['userSelect'], FILTER_VALIDATE_INT);
+					$uID = filter_var($data['userSelect'], FILTER_VALIDATE_INT);
 					$query = "UPDATE `users` SET `userName`= '$user', `userType`= '$userLevel' WHERE `userID` = $uID;";
 					if(db_query($query) == 1) {
 						if(filter_var($user, FILTER_VALIDATE_EMAIL)) {
@@ -545,12 +559,10 @@ if(isset($_POST['action'])) {
 					} else {
 						$response = "User change failed";
 					}
-					header("Content-Type: application/json");
-					die(json_encode($response));
 					break;
 				case 3:
 					e_log(8,"Delete user $user");
-					$uID = filter_var($_POST['userSelect'], FILTER_VALIDATE_INT);
+					$uID = filter_var($data['userSelect'], FILTER_VALIDATE_INT);
 					$query = "DELETE FROM `users` WHERE `userID` = $uID;";
 					if(db_query($query) == 1) {
 						if(filter_var($user, FILTER_VALIDATE_EMAIL)) {
@@ -563,33 +575,31 @@ if(isset($_POST['action'])) {
 					} else {
 						$response = "Delete user failed";
 					}
-					header("Content-Type: application/json");
-					die(json_encode($response));
 					break;
 				default:
-					$message = "Unknown action for managing users";
-					e_log(1,$message);
-					die($message);
+					$response = "Unknown action for managing users";
+					e_log(1,$response);
 			}
+			sendJSONResponse($response);
 			break;
 		case "mlog":
 			if($_SESSION['sud']['userType'] > 1) {
 			    $lfile = is_dir(CONFIG['logfile']) ? CONFIG['logfile'].'/syncmarks.log':CONFIG['logfile'];
-				die(file_get_contents($lfile));
+				sendJSONResponse(file_get_contents($lfile));
 			} else {
 				$message = "Not allowed to read server logfile.";
 				e_log(2,$message);
-				die($message);
+				sendJSONResponse($message);
 			}
 			break;
 		case "mrefresh":
 			if($_SESSION['sud']['userType'] > 1) {	
 			    $lfile = is_dir(CONFIG['logfile']) ? CONFIG['logfile'].'/syncmarks.log':CONFIG['logfile'];
-				die(file_get_contents($lfile));
+				sendJSONResponse(file_get_contents($lfile));
 			} else {
 				$message = "Not allowed to read server logfile.";
 				e_log(2,$message);
-				die($message);
+				sendJSONResponse($message);
 			}
 			break;
 		case "mclear":
@@ -597,55 +607,25 @@ if(isset($_POST['action'])) {
 			if($_SESSION['sud']['userType'] > 1) {
 				$lfile = is_dir(CONFIG['logfile']) ? CONFIG['logfile'].'/syncmarks.log':CONFIGg['logfile'];
 				file_put_contents($lfile,"");
-				die(file_get_contents($lfile));
+				sendJSONResponse(file_get_contents($lfile));
 			}
-				
 			die();
 			break;
-		case "madd":
-			$bmParentID = filter_var($_POST['folder'], FILTER_SANITIZE_STRING);
-			$bmURL = validate_url(trim($_POST['url']));
-			e_log(8,"Try to add manually new bookmark ".$bmURL);
-			
-			if(strpos($bmURL,'http') != 0) {
-				e_log(1,"Given string is not a real URL, cant add this.");
-				exit;
-			}
-			
-			$bookmark['url'] = $bmURL;
-			$bookmark['folder'] = $bmParentID;
-			$bookmark['title'] = getSiteTitle($bmURL);
-			$bookmark['id'] = unique_code(12);
-			$bookmark['type'] = 'bookmark';
-			$bookmark['added'] = round(microtime(true) * 1000);
-			
-			$res = addBookmark($bookmark);
-			
-			if($res === 1) {
-				if(!isset($_POST['rc'])) {
-					e_log(8,"Manually added bookmark.");
-					die(bmTree());
-				} else {
-					die(e_log(8,"Roundcube added bookmark."));
-				}
-			} else {
-				echo $res;
-				http_response_code(417);
-			}
-			
-			break;
 		case "mdel":
-			$bmID = filter_var($_POST['id'], FILTER_SANITIZE_STRING);
+			$bmID = json_decode($_POST['data'], true);
 			$delMark = delMark($bmID);
 			if($delMark != 0) {
-				if(!isset($_POST['rc'])) {
+				if(!isset($_POST['add'])) {
 					e_log(8,"Bookmark $bmID removed");
-					die();
+					sendJSONResponse($bmID);
 				} else {
-					die(e_log(8,"Bookmark $bmID deleted by Roundcube"));
+					e_log(8,"Bookmark $bmID deleted by Roundcube");
+					sendJSONResponse($bmID);
 				}
 			} else {
-				die(e_log(2,"There was an problem removing the bookmark, please check the logfile"));
+				$message = "There was an problem removing the bookmark, please check the logfile";
+				e_log(2, $message);
+				sendJSONResponse($message);
 			}
 			break;
 		case "logout":
@@ -663,6 +643,7 @@ if(isset($_POST['action'])) {
 				</div>";
 				echo htmlFooter();
 			}
+			session_destroy();
 			exit;
 			break;
 		case "pbupdate":
@@ -776,664 +757,6 @@ if(isset($_POST['action'])) {
 			} else {
 				die(json_encode('Editing users not allowed'));
 			}
-			break;
-		default:
-			die(e_log(1, "Unknown Action"));
-	}
-}
-
-if(isset($_POST['caction'])) {
-	switch($_POST['caction']) {
-		case "addmark":
-			$bookmark = json_decode($_POST['bookmark'], true);
-			e_log(8,"Try to add new bookmark '".$bookmark['title']."'");
-			$stime = round(microtime(true) * 1000);
-			$bookmark['added'] = $stime;
-			$bookmark['title'] = htmlspecialchars(mb_convert_encoding(htmlspecialchars_decode($bookmark['title'], ENT_QUOTES),"UTF-8"),ENT_QUOTES,'UTF-8', false);
-			e_log(9, print_r($bookmark, true));
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-			if(array_key_exists('url',$bookmark)) $bookmark['url'] = validate_url($bookmark['url']);
-			if(strtolower(getClientType($_SERVER['HTTP_USER_AGENT'])) != "firefox") $bookmark = cfolderMatching($bookmark);
-			$stime = (!isset($_POST['s'])) ? 0:$stime;
-			header("Content-Type: application/json");
-			if($bookmark['type'] == 'bookmark' && isset($bookmark['url'])) {
-				$response = json_encode(addBookmark($bookmark));
-				updateClient($client, strtolower(getClientType($_SERVER['HTTP_USER_AGENT'])), $stime, true);
-				die($response);
-			} else if($bookmark['type'] == 'folder') {
-				$response = addFolder($bookmark);
-				updateClient($client, strtolower(getClientType($_SERVER['HTTP_USER_AGENT'])), $stime, true);
-				die(json_encode($response));
-			} else {
-				$message = "This bookmark is not added, some parameters are missing";
-				e_log(1, $message);
-				die(json_encode($message));
-			}
-			break;
-		case "movemark":
-			$bookmark = json_decode($_POST['bookmark'],true);
-			if(CONFIG['cexp'] == true && CONFIG['loglevel'] == 9) {
-				$filename = is_dir(CONFIG['logfile']) ? CONFIG['logfile']."/movemark_".time().".json":"movemark_".time().".json";
-				e_log(8,"Write move bookmark json to $filename");
-				file_put_contents($filename,json_encode($bookmark),true);
-			}
-
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-			$ctime = round(microtime(true) * 1000);
-			$response = json_encode(moveBookmark($bookmark));
-			$ctime = (filter_var($_POST['s'], FILTER_SANITIZE_STRING) === 'false') ? 0:$ctime;
-			updateClient($client, strtolower(getClientType($_SERVER['HTTP_USER_AGENT'])), $ctime, true);
-			die($response);
-			break;
-		case "editmark":
-			$bookmark = json_decode(rawurldecode($_POST['bookmark']),true);
-			(array_key_exists('url',$bookmark)) ? die(editBookmark($bookmark)) : die(editFolder($bookmark));
-			break;
-		case "delmark":
-			$bookmark = json_decode(rawurldecode($_POST['bookmark']),true);
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-			$ctime = round(microtime(true) * 1000);
-			e_log(8,"Try to identify bookmark to delete");
-			if(isset($bookmark['url'])) {
-				$url = prepare_url($bookmark['url']);
-				$query = "SELECT `bmID` FROM `bookmarks` WHERE `bmType` = 'bookmark' AND `bmURL` = '$url' AND `userID` = ".$_SESSION['sud']['userID'].";";
-			} else {
-				$query = "SELECT `bmID` FROM `bookmarks` WHERE `bmType` = 'folder' AND `bmTitle` = '".$bookmark['title']."' AND `userID` = ".$_SESSION['sud']['userID'].";";
-			}
-
-			$bData = db_query($query);
-			header("Content-Type: application/json");
-			if(count($bData) == 1) {
-				e_log(2, "Bookmark found, trying to remove it");
-				$delmark = delMark($bData[0]['bmID']);
-				die(json_encode($delmark));
-			} else if (count($bData) > 1) {
-				$message = "No unique bookmark found, doing nothing";
-				e_log(2,$message);
-				die(json_encode($message));
-			} else {
-				e_log(2, "Bookmark not found, mark as deleted");
-				die(json_encode(1));
-			}
-			break;
-		case "startup":
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-			$ctype = getClientType($_SERVER['HTTP_USER_AGENT']);
-			$ctime = (filter_var($_POST['s'], FILTER_SANITIZE_STRING) === 'false') ? 0:round(microtime(true) * 1000);
-			$changes = getChanges($client, $ctype, $ctime);
-
-			if(CONFIG['cexp'] == true && CONFIG['loglevel'] == 9) {	
-				$filename = is_dir(CONFIG['logfile']) ? CONFIG['logfile']."/startup_".time().".json":"startup_".time().".json";
-				e_log(8,"Write startup json to $filename");
-				file_put_contents($filename,json_encode($changes),true);
-			}
-			header("Content-Type: application/json");
-			die(json_encode($changes,JSON_UNESCAPED_SLASHES));
-			break;
-		case "cfolder":
-			$ctime = round(microtime(true) * 1000);
-			$fname = filter_var($_POST['fname'], FILTER_SANITIZE_STRING);
-			$fbid = filter_var($_POST['fbid'], FILTER_SANITIZE_STRING);
-			die(cfolder($ctime,$fname,$fbid));
-			break;
-		case "import":
-			$jmarks = json_decode($_POST['bookmark'],true);
-			$jerrmsg = "";
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-			$partial = (isset($_POST['p'])) ? filter_var($_POST['p'], FILTER_VALIDATE_INT):0;
-			switch (json_last_error()) {
-				case JSON_ERROR_NONE:
-					$jerrmsg = '';
-				break;
-				case JSON_ERROR_DEPTH:
-					$jerrmsg = 'Maximum stack depth exceeded';
-				break;
-				case JSON_ERROR_STATE_MISMATCH:
-					$jerrmsg = 'Underflow or the modes mismatch';
-				break;
-				case JSON_ERROR_CTRL_CHAR:
-					$jerrmsg = 'Unexpected control character found';
-				break;
-				case JSON_ERROR_SYNTAX:
-					$jerrmsg = 'Syntax error, malformed JSON';
-				break;
-				case JSON_ERROR_UTF8:
-					$jerrmsg = 'Malformed UTF-8 characters, possibly incorrectly encoded';
-				break;
-				default:
-					$jerrmsg = 'Unknown error';
-			}
-			
-			if(strlen($jerrmsg) > 0) {
-				e_log(1,"JSON error: ".$jerrmsg);
-				$filename = is_dir(CONFIG['logfile']) ? CONFIG['logfile']."/import_".time().".json":"import_".time().".json";
-				e_log(8,"JSON file written as $filename");
-				file_put_contents($filename,urldecode($_POST['bookmark']),true);
-				header("Content-Type: application/json");
-				die(json_encode($jerrmsg));
-			}
-
-			$client = $_POST['client'];
-			$ctype = getClientType($_SERVER['HTTP_USER_AGENT']);
-			$ctime = round(microtime(true) * 1000);
-			
-			if($partial === 0) delUsermarks($_SESSION['sud']['userID']);
-			
-			$armarks = parseJSON($jmarks);
-			$ctime = (filter_var($_POST['s'], FILTER_SANITIZE_STRING) === 'false') ? 0:$ctime;
-			updateClient($client, $ctype, $ctime, true);
-			header("Content-Type: application/json");
-			die(json_encode(importMarks($armarks,$_SESSION['sud']['userID'])));
-			break;
-		case "getpurl":
-			$url = validate_url($_POST['url']);
-			e_log(8,"Received new pushed URL: ".$url);
-			$target = (isset($_POST['tg'])) ? filter_var($_POST['tg'], FILTER_SANITIZE_STRING) : '0';
-			if(newNotification($url, $target) !== 0) die("URL successful pushed.");
-			break;
-		case "rmessage":
-			$message = isset($_POST['message']) ? filter_var($_POST['message'], FILTER_VALIDATE_INT):0;
-			$loop = filter_var($_POST['lp'], FILTER_SANITIZE_STRING) == 'aNoti' ? 1 : 0;
-
-			if($message > 0) {
-				e_log(8,"Try to delete notification $message");
-				$query = "DELETE FROM `notifications` WHERE `userID` = ".$_SESSION['sud']['userID']." AND `id` = $message;";
-				$count = db_query($query);
-				($count === 1) ? e_log(8,"Notification successfully removed") : e_log(9,"Error, removing notification");
-			}
-			
-			die(notiList($_SESSION['sud']['userID'], $loop));
-			break;
-		case "soption":
-			$option = filter_var($_POST['option'], FILTER_SANITIZE_STRING);
-			$value = filter_var(filter_var($_POST['value'], FILTER_SANITIZE_NUMBER_INT), FILTER_VALIDATE_INT);
-			e_log(8,"Option received: ".$option.":".$value);
-			$oOptionsA = json_decode($_SESSION['sud']['uOptions'],true);
-			$oOptionsA[$option] = $value;
-			$query = "UPDATE `users` SET `uOptions`='".json_encode($oOptionsA)."' WHERE `userID`=".$_SESSION['sud']['userID'].";";
-			header("Content-Type: application/json");
-			if(db_query($query) !== false) {
-				e_log(8,"Option saved");
-				die(json_encode(true));
-			} else {
-				e_log(9,"Error, saving option");
-				die(json_encode(false));
-			}
-			break;
-		case "getclients":
-			e_log(8,"Try to get list of clients");
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-			$query = "SELECT `cid`, IFNULL(`cname`, `cid`) `cname`, `ctype`, `lastseen` FROM `clients` WHERE `uid` = ".$_SESSION['sud']['userID']." AND NOT `cid` = '$client';";
-			$clientList = db_query($query);
-			e_log(8,"Found ".count($clientList)." clients. Send list to '$client'.");
-			
-			uasort($clientList, function($a, $b) {
-				return strnatcasecmp($a['cname'], $b['cname']);
-			});
-			
-			if (!empty($clientList)) {
-				foreach($clientList as $key => $clients) {
-					$myObj[$key]['id'] =	$clients['cid'];
-					$myObj[$key]['name'] = 	$clients['cname'];
-					$myObj[$key]['type'] = 	$clients['ctype'];
-					$myObj[$key]['date'] = 	$clients['lastseen'];
-				}
-				$all = array('id'=>'0', 'name'=>'All Clients', 'type'=>'', 'date'=>'');
-				array_unshift($myObj, $all);
-			} else {
-				$myObj[0]['id'] =	'0';
-				$myObj[0]['name'] =	'All Clients';
-				$myObj[0]['type'] =	'';
-				$myObj[0]['date'] =	'';
-			}
-			
-			if(CONFIG['cexp'] == true && CONFIG['loglevel'] == 9) {
-				$filename = is_dir(CONFIG['logfile']) ? CONFIG['logfile']."/clist_".time().".json":"clist_".time().".json";
-				e_log(8,"Write clientlist to $filename");
-				file_put_contents($filename,json_encode($myObj),true);
-			}
-			
-			header("Content-Type: application/json");
-			die(json_encode($myObj));
-			break;
-		case "tlg":
-			$userID = $_SESSION['sud']['userID'];
-			$query = "SELECT * FROM `c_token` WHERE `userID` = $userID;";
-			$tData = db_query($query);
-			header("Content-Type: application/json");
-			die(json_encode($tData));
-			break;
-		case "tld":
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-			$userID = $_SESSION['sud']['userID'];
-			$query = "DELETE FROM `c_token` WHERE `cid` = '$client' AND `userID` = $userID;";
-			$tData = db_query($query);
-			header("Content-Type: application/json");
-			die(json_encode($tData));
-			break;
-		case "tl":
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-			e_log(8,"Token request from client '$client'");
-			$type = getClientType($_SERVER['HTTP_USER_AGENT']);
-			$time = round(microtime(true) * 1000);
-			$ctime = (filter_var($_POST['s'], FILTER_SANITIZE_STRING) === 'false') ? 0:$time;
-			$tResponse['message'] = updateClient($client, $type, $ctime);
-			$userID = $_SESSION['sud']['userID'];
-			$query = "SELECT `c_token`.*, `clients`.`cname` FROM `c_token` INNER JOIN `clients` ON `clients`.`cid` = `c_token`.`cid` WHERE `c_token`.`cid` = '$client' AND `userID` = $userID;";
-			$tData = db_query($query);
-			$expireTime = time()+60*60*24*7;
-			$token = bin2hex(openssl_random_pseudo_bytes(32));
-			$thash = password_hash($token, PASSWORD_DEFAULT);
-			if(count($tData) > 0) {
-				$query = "UPDATE `c_token` SET `tHash` = '$thash', `exDate` = '$expireTime' WHERE `cid` = '$client' AND `userID` = $userID;";
-				$tResponse['cname'] = $tData[0]['cname'];
-			} else {
-				$query = "INSERT INTO `c_token` (`cid`, `tHash`, `exDate`, `userID`) VALUES ('$client', '$thash', '$expireTime', $userID);";
-				$tResponse['cname'] = '';
-			}
-			db_query($query);
-			$tResponse['token'] = $token;
-			
-			e_log(8, "Logout $client");
-			unset($_SESSION['sauth']);
-			session_destroy();
-
-			header("Content-Type: application/json");
-			e_log(8, "Send new token to client $client");
-			die(json_encode($tResponse));
-			break;
-		case "cinfo":
-			e_log(8,"Request client info");
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-			$query = "SELECT `cname`, `ctype`, `lastseen` FROM `clients` WHERE `cid` = '$client' AND `uid` = ".$_SESSION['sud']['userID'].";";
-			$clientData = db_query($query);
-			if(count($clientData) > 0) {
-				e_log(8,"Send client info to '$client'");
-			} else {
-				e_log(2,"Client not found.");
-				$clientData[0]['lastseen'] = 0;
-				$clientData[0]['cname'] = '';
-				$clientData[0]['ctype'] = '';
-			}
-			
-			header("Content-Type: application/json");
-			die(json_encode($clientData[0]));
-			break;
-		case "cfsync":
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-			$query = "SELECT `fs`, `lastseen` FROM `clients` WHERE `cid` = '$client';";
-			$fsdata = db_query($query)['0'];
-			header("Content-Type: application/json");
-			die(json_encode($fsdata));
-			break;
-		case "gurls":
-			$client = (isset($_POST['client'])) ? filter_var($_POST['client'], FILTER_SANITIZE_STRING) : '0';
-			e_log(8,"Request pushed sites for '$client'");
-			$query = "SELECT * FROM `notifications` WHERE `nloop` = 1 AND `userID` = ".$_SESSION['sud']['userID']." AND `client` IN ('".$client."','0');";
-			$uOptions = json_decode($_SESSION['sud']['uOptions'],true);
-			$notificationData = db_query($query);
-			if (!empty($notificationData)) {
-				e_log(8,"Found ".count($notificationData)." links. Will push them to the client.");
-				foreach($notificationData as $key => $notification) {
-					$myObj[$key]['title'] = html_entity_decode($notification['title'], ENT_QUOTES | ENT_XML1, 'UTF-8');
-					$myObj[$key]['url'] = $notification['message'];
-					$myObj[$key]['nkey'] = $notification['id'];
-					$myObj[$key]['nOption'] = $uOptions['notifications'];
-				}
-				header("Content-Type: application/json");
-				die(json_encode($myObj));
-			} else {
-				e_log(8,"No pushed sites found");
-				header("Content-Type: application/json");
-				die(json_encode("0"));
-			}
-			break;
-		case "durl":
-			e_log(8,"Hide notification");
-			$notification = filter_var($_POST['durl'], FILTER_VALIDATE_INT);
-			$query = "UPDATE `notifications` SET `nloop`= 0, `ntime`= '".time()."' WHERE `id` = $notification AND `userID` = ".$_SESSION['sud']['userID'];
-			die(db_query($query));
-			break;
-		case "bmedt":
-			$title = filter_var($_POST['title'], FILTER_SANITIZE_STRING);
-			$id = filter_var($_POST['id'], FILTER_SANITIZE_STRING);
-			e_log(8,"Edit entry '$title'");
-			$url = (isset($_POST['url']) && strlen($_POST['url']) > 4) ? '\''.validate_url($_POST['url']).'\'' : 'NULL';
-			$query = "UPDATE `bookmarks` SET `bmTitle` = '$title', `bmURL` = $url, `bmAdded` = '".round(microtime(true) * 1000)."' WHERE `bmID` = '$id' AND `userID` = ".$_SESSION['sud']['userID'].";";
-			$count = db_query($query);
-			($count > 0) ? die(true) : die(false);
-			break;
-		case "bmmv":
-			$id = filter_var($_POST['id'], FILTER_SANITIZE_STRING);
-			e_log(8,"Move bookmark $id");
-			$folder = filter_var($_POST['folder'], FILTER_SANITIZE_STRING);
-			$query = "SELECT MAX(bmIndex)+1 AS 'index' FROM `bookmarks` WHERE `bmParentID` = '$folder';";
-			$folderData = db_query($query);
-			$query = "UPDATE `bookmarks` SET `bmIndex` = ".$folderData[0]['index'].", `bmParentID` = '$folder', `bmAdded` = '".round(microtime(true) * 1000)."' WHERE `bmID` = '$id' AND `userID` = ".$_SESSION['sud']['userID'].";";
-			$count = db_query($query);
-			($count > 0) ? die(true) : die(false);
-			break;
-		case "arename":
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-			$name = filter_var($_POST['nname'], FILTER_SANITIZE_STRING);
-			e_log(8,"Rename client $client to $name");
-			$query = "UPDATE `clients` SET `cname` = '".$name."' WHERE `uid` = ".$_SESSION['sud']['userID']." AND `cid` = '".$client."';";
-			$count = db_query($query);
-			($count > 0) ? die(bClientlist($_SESSION['sud']['userID'])) : die(false);
-			break;
-		case "adel":
-			$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-			e_log(8,"Delete client $client");
-			$query = "DELETE FROM `clients` WHERE `uid` = ".$_SESSION['sud']['userID']." AND `cid` = '$client';";
-			$count = db_query($query);
-			($count > 0) ? die(bClientlist($_SESSION['sud']['userID'])) : die(false);
-			break;
-		case "cmail":
-			e_log(8,"Change e-mail for ".$_SESSION['sud']['userName']);
-			$nmail = filter_var($_POST['mail'],FILTER_SANITIZE_EMAIL);
-			header("Content-Type: application/json");
-			if(filter_var($nmail, FILTER_VALIDATE_EMAIL)) {
-				$query = "UPDATE `users` SET `userMail` = '$nmail' WHERE `userID` = ".$_SESSION['sud']['userID'].";";
-				die(json_encode(db_query($query)));
-			} else {
-				e_log(1,"No valid E-Mail. Stop changing E-Mail");
-				die(json_encode("No valid mail address. Mail not changed."));
-			}
-			die();
-			break;
-		case "muedt":
-			if($_SESSION['sud']['userType'] < 2) {	
-				e_log(1,"Stop user change, no sufficient privileges.");
-				die();
-			}
-			$del = false;
-			$headers = "From: SyncMarks <".CONFIG['sender'].">";
-			$url = $_SERVER['REQUEST_SCHEME']."://".$_SERVER['HTTP_HOST'].$_SERVER['SCRIPT_NAME'];
-			$variant = filter_var($_POST['type'], FILTER_VALIDATE_INT);
-			$password = (isset($_POST['p']) && $_POST['p'] != '') ? filter_var($_POST['p'], FILTER_SANITIZE_STRING):gpwd(16);
-			$userLevel = filter_var($_POST['userLevel'], FILTER_VALIDATE_INT);
-			$user = filter_var($_POST['nuser'], FILTER_SANITIZE_STRING);
-			$mail = filter_var($user, FILTER_VALIDATE_EMAIL) ? $user:null;
-
-			switch($variant) {
-				case 1:
-					$pwd = password_hash($password,PASSWORD_DEFAULT);
-					e_log(8,"Try to add new user $user");
-					$query = "INSERT INTO `users` (`userName`,`userMail`,`userType`,`userHash`) VALUES ('$user', NULLIF('$mail',''), '$userLevel', '$pwd')";
-					$nuid = db_query($query);
-					if($nuid > 0) {
-						if(filter_var($mail, FILTER_VALIDATE_EMAIL)) {
-							$response = $nuid;
-							$message = "Hello,\r\na new account with the following credentials is created and stored encrypted on for SyncMarks:\r\nUsername: $user\r\nPassword: $password\r\n\r\nYou can login at $url";
-							if(!mail ($mail, "Account created",$message,$headers)) {
-								e_log(1,"Error sending data for created user account to user");
-								$response = "User created successful, E-Mail could not send";
-							}
-						} else {
-							$response = $nuid;
-						}
-						$bmAdded = round(microtime(true) * 1000);
-						$query = "INSERT INTO `bookmarks` (`bmID`,`bmParentID`,`bmIndex`,`bmTitle`,`bmType`,`bmURL`,`bmAdded`,`userID`) VALUES ('unfiled_____', 'root________', 0, 'Other Bookmarks', 'folder', NULL, ".$bmAdded.", $nuid)";
-						db_query($query);
-						$query = "INSERT INTO `bookmarks` (`bmID`,`bmParentID`,`bmIndex`,`bmTitle`,`bmType`,`bmURL`,`bmAdded`,`userID`) VALUES ('".unique_code(12)."', 'unfiled_____', 0, 'Git Repository', 'bookmark', 'https://codeberg.org/Offerel', ".$bmAdded.", $nuid)";
-						db_query($query);
-					} else {
-						$response = "User creation failed";
-					}
-					header("Content-Type: application/json");
-					die(json_encode($response));
-					break;
-				case 2:
-					e_log(8,"Updating user $user");
-					$uID = filter_var($_POST['userSelect'], FILTER_VALIDATE_INT);
-					$query = "UPDATE `users` SET `userName`= '$user', `userType`= '$userLevel' WHERE `userID` = $uID;";
-					if(db_query($query) == 1) {
-						if(filter_var($user, FILTER_VALIDATE_EMAIL)) {
-							$response = "User changed successful, Try to send E-Mail to user";
-							$message = "Hello,\r\nyour account is changed for SyncMarks. You can login at $url";
-							if(!mail ($user, "Account changed",$message,$headers)) e_log(1,"Error sending email for changed user account");
-						} else {
-							$response = "User changed successful, No mail send to user";
-						}
-					} else {
-						$response = "User change failed";
-					}
-					header("Content-Type: application/json");
-					die(json_encode($response));
-					break;
-				case 3:
-					e_log(8,"Delete user $user");
-					$uID = filter_var($_POST['userSelect'], FILTER_VALIDATE_INT);
-					$query = "DELETE FROM `users` WHERE `userID` = $uID;";
-					if(db_query($query) == 1) {
-						if(filter_var($user, FILTER_VALIDATE_EMAIL)) {
-							$response = "User deleted, Try to send E-Mail to user";
-							$message = "Hello,\r\nyour account '$user' and all it's data is removed from $url.";
-							if(!mail ($user, "Account removed",$message,$headers)) e_log(1,"Error sending data for created user account to user");
-						} else {
-							$response = "User deleted successful, No mail send to user";
-						}
-					} else {
-						$response = "Delete user failed";
-					}
-					header("Content-Type: application/json");
-					die(json_encode($response));
-					break;
-				default:
-					$message = "Unknown action for managing users";
-					e_log(1,$message);
-					die($message);
-			}
-			break;
-		case "mlog":
-			if($_SESSION['sud']['userType'] > 1) {
-			    $lfile = is_dir(CONFIG['logfile']) ? CONFIG['logfile'].'/syncmarks.log':CONFIG['logfile'];
-				die(file_get_contents($lfile));
-			} else {
-				$message = "Not allowed to read server logfile.";
-				e_log(2,$message);
-				die($message);
-			}
-			break;
-		case "mrefresh":
-			if($_SESSION['sud']['userType'] > 1) {	
-			    $lfile = is_dir(CONFIG['logfile']) ? CONFIG['logfile'].'/syncmarks.log':CONFIG['logfile'];
-				die(file_get_contents($lfile));
-			} else {
-				$message = "Not allowed to read server logfile.";
-				e_log(2,$message);
-				die($message);
-			}
-			break;
-		case "mclear":
-			e_log(8,"Clear logfile");
-			if($_SESSION['sud']['userType'] > 1) {
-				$lfile = is_dir(CONFIG['logfile']) ? CONFIG['logfile'].'/syncmarks.log':CONFIGg['logfile'];
-				file_put_contents($lfile,"");
-				die(file_get_contents($lfile));
-			}
-				
-			die();
-			break;
-		case "madd":
-			$bmParentID = filter_var($_POST['folder'], FILTER_SANITIZE_STRING);
-			$bmURL = validate_url(trim($_POST['url']));
-			e_log(8,"Try to add manually new bookmark ".$bmURL);
-			
-			if(strpos($bmURL,'http') != 0) {
-				e_log(1,"Given string is not a real URL, cant add this.");
-				exit;
-			}
-			
-			$bookmark['url'] = $bmURL;
-			$bookmark['folder'] = $bmParentID;
-			$bookmark['title'] = getSiteTitle($bmURL);
-			$bookmark['id'] = unique_code(12);
-			$bookmark['type'] = 'bookmark';
-			$bookmark['added'] = round(microtime(true) * 1000);
-			
-			$res = addBookmark($bookmark);
-			
-			if($res === 1) {
-				if(!isset($_POST['rc'])) {
-					e_log(8,"Manually added bookmark.");
-					die(bmTree());
-				} else {
-					die(e_log(8,"Roundcube added bookmark."));
-				}
-			} else {
-				echo $res;
-				http_response_code(417);
-			}
-			
-			break;
-		case "mdel":
-			$bmID = filter_var($_POST['id'], FILTER_SANITIZE_STRING);
-			$delMark = delMark($bmID);
-			if($delMark != 0) {
-				if(!isset($_POST['rc'])) {
-					e_log(8,"Bookmark $bmID removed");
-					die();
-				} else {
-					die(e_log(8,"Bookmark $bmID deleted by Roundcube"));
-				}
-			} else {
-				die(e_log(2,"There was an problem removing the bookmark, please check the logfile"));
-			}
-			break;
-		case "pupdate":
-			e_log(8,"User change: Updating user password started");
-			$opassword = filter_var($_POST['opassword'], FILTER_SANITIZE_STRING);
-			$npassword = filter_var($_POST['npassword'], FILTER_SANITIZE_STRING);
-			$cpassword = filter_var($_POST['cpassword'], FILTER_SANITIZE_STRING);
-
-			if($opassword != "" && $npassword !="" && $cpassword !="") {
-				e_log(8,"User change: Data complete entered");
-				if(password_verify($opassword,$_SESSION['sud']['userHash'])) {
-					e_log(8,"User change: Verify original password");
-					if($npassword === $cpassword) {
-						e_log(8,"User change: New and confirmed password");
-						if($npassword != $opassword) {
-							$password = password_hash($npassword,PASSWORD_DEFAULT);
-							$query = "UPDATE `users` SET `userHash`='$password' WHERE `userID`=".$_SESSION['sud']['userID'].";";
-							db_query($query);
-							e_log(8,"User change: Password changed");
-						} else {
-							e_log(2,"User change: Old and new password identical, user not changed");
-						}
-					}
-				} else {
-					e_log(2,"User change: Old password mismatch");
-				}
-			} else {
-				e_log(2,"User change: Data missing, process failed");
-			}
-
-			unset($_SESSION['sauth']);
-			e_log(8,"User logged out");
-			echo htmlHeader();
-			echo "<div id='loginbody'>
-				<div id='loginform'>
-					<div id='loginformh'>Logout successful</div>
-					<div id='loginformt'>User logged out. <a href='".$_SERVER['SCRIPT_NAME']."'>Login</a> again</div>
-				</div>
-			</div>";
-			echo htmlFooter();
-
-			die();
-			break;
-		case "pbupdate":
-			e_log(8,"Pushbullet: Updating Pushbullet information.");
-			$password = filter_var($_POST['password'], FILTER_SANITIZE_STRING);
-			$ptoken = filter_var($_POST['ptoken'], FILTER_SANITIZE_STRING);
-			$pdevice = filter_var($_POST['pdevice'], FILTER_SANITIZE_STRING);
-			$pbe = filter_var($_POST['pbe'], FILTER_SANITIZE_STRING);
-
-			if(password_verify($password,$_SESSION['sud']['userHash'])) {
-				$token = edcrpt('en', $ptoken);
-				$device = edcrpt('en', $pdevice);
-				$pbEnable = filter_var($pbe,FILTER_VALIDATE_BOOLEAN) ? '1' : '0';
-		
-				$oOptionsA = json_decode($_SESSION['sud']['uOptions'],true);
-				$oOptionsA['pAPI'] = $token;
-				$oOptionsA['pDevice'] = $device;
-				$oOptionsA['pbEnable'] = $pbEnable;
-		
-				$query = "UPDATE `users` SET `uOptions`='".json_encode($oOptionsA)."' WHERE `userID`=".$_SESSION['sud']['userID'].";";
-				$count = db_query($query);
-				($count === 1) ? e_log(8,"Option saved") : e_log(9,"Error, saving option");
-				header("location: ?");
-				die();
-			}
-			else {
-				e_log(1,"Password mismatch. Pushbullet not updated.");
-				die("Password mismatch. Pushbullet not updated.");
-			}
-			die();
-			break;
-		case "uupdate":
-			e_log(8,"User change: Updating user name started");
-			$opassword = filter_var($_POST['opassword'], FILTER_SANITIZE_STRING);
-			$username = filter_var($_POST['username'], FILTER_SANITIZE_STRING);
-
-			if($opassword != "") {
-				e_log(8,"User change: Data complete entered");
-				if(password_verify($opassword, $_SESSION['sud']['userHash'])) {
-					e_log(8,"User change: Verify original password");
-					$query = "UPDATE `users` SET `userName`='$username' WHERE `userID`=".$_SESSION['sud']['userID'].";";
-					db_query($query);
-					e_log(8,"User change: Username changed");
-				}
-				else {
-					e_log(2,"User change: Failed to verify original password");
-				}
-			}
-			else {
-				e_log(2,"User change: Data missing");
-			}
-			unset($_SESSION['sauth']);
-			e_log(8,"User logged out");
-			echo htmlHeader();
-			echo "<div id='loginbody'>
-				<div id='loginform'>
-					<div id='loginformh'>Logout successful</div>
-					<div id='loginformt'>User logged out. <a href='?'>Login</a> again</div>
-				</div>
-			</div>";
-			echo htmlFooter();
-			die();
-			break;
-		case "export":
-			e_log(8,"Request bookmark export");
-			$ctype = getClientType($_SERVER['HTTP_USER_AGENT']);
-			$ctime = round(microtime(true) * 1000);
-			$format = filter_var($_POST['type'], FILTER_SANITIZE_STRING);
-			switch($format) {
-				case "html":
-					e_log(8,"Exporting in HTML format for download");
-					die(html_export());
-					break;
-				case "json":
-					e_log(8,"Exporting in JSON format");
-					$client = filter_var($_POST['client'], FILTER_SANITIZE_STRING);
-					$bookmarks = json_encode(getBookmarks());
-					if(CONFIG['loglevel'] == 9 && CONFIG['cexp'] == true) {
-						$filename = is_dir(CONFIG['logfile']) ? CONFIG['logfile']."/export_".time().".json":"export_".time().".json";
-						file_put_contents($filename,$bookmarks,true);
-						e_log(8,"Export file is saved to $filename");
-					}
-					$bcount = count(json_decode($bookmarks));
-					e_log(8,"Send $bcount bookmarks to '$client'");
-					$ctime = (filter_var($_POST['s'], FILTER_SANITIZE_STRING) === 'false') ? 0:$ctime;
-					updateClient($client, $ctype, $ctime, true);
-
-					header("Content-Type: application/json");
-					die($bookmarks);
-					break;
-				default:
-					die(e_log(2,"Unknown export format, exit process"));
-			}
-			exit;
 			break;
 		case "checkdups":
 			e_log(8,"Checking for duplicated bookmarks by url");
@@ -1447,47 +770,11 @@ if(isset($_POST['caction'])) {
 				}
 				$dubData[$key]['subs'] = $subData;
 			}
-			header("Content-Type: application/json");
-			die(json_encode($dubData));
-			break;
-		case "logout":
-			e_log(8,"Logout user ".$_SESSION['sauth']);
-			unset($_SESSION['sauth']);
-			clearAuthCookie();
-			e_log(8,"User logged out");
-			if(!isset($_POST['client'])) {
-				echo htmlHeader();
-				echo "<div id='loginbody'>
-					<div id='loginform'>
-						<div id='loginformh'>Logout successful</div>
-						<div id='loginformt'>User logged out. <a href='?'>Login</a> again</div>
-					</div>
-				</div>";
-				echo htmlFooter();
-			}
-			exit;
-			break;
-		case "maddon":
-			$rResponse['bookmarks'] = showBookmarks(1);
-			$rResponse['folders'] = getUserFolders($_SESSION['sud']['userID']);
-			header("Content-Type: application/json");
-			die(json_encode($rResponse));
-			break;
-		case "getUsers":
-			header("Content-Type: application/json");
-			if($_SESSION['sud']['userType'] == 2) {
-				$query = "SELECT `userID`, `userName`, `userType` FROM `users` ORDER BY `userName`;";
-				$uData = db_query($query);
-				die(json_encode($uData));
-			} else {
-				die(json_encode('Editing users not allowed'));
-			}
+			sendJSONResponse($dubData);
 			break;
 		default:
-			header("Content-Type: application/json");
-			die(json_encode("Unknown cAction"));
+			die(e_log(1, "Unknown Action"));
 	}
-	exit;
 }
 
 if(isset($_GET['link'])) {
@@ -1566,7 +853,7 @@ function newNotification($url, $target) {
 	$options = json_decode($_SESSION['sud']['uOptions'],true);
 	
 	if(strlen($options['pAPI']) > 1 && strlen($options['pDevice']) > 1 && $options['pbEnable'] == "1") {
-		pushlink($title,$url);
+		pushlink($title, $url);
 	} else {
 		e_log(2,"Can't send to Pushbullet, missing data. Please check options");
 	}
@@ -1597,29 +884,35 @@ function fWay($parent, $user, $str) {
 
 function delMark($bmID) {
 	$count = 0;
-	e_log(8,"Delete bookmark '$bmID'");
+	$bms = implode(", ", $bmID);
+	e_log(8,"Delete bookmark(s) $bms");
 
-	$query = "SELECT `bmParentID`, `bmIndex`, `bmURL` FROM `bookmarks` WHERE `bmID` = '$bmID' AND `userID` = ".$_SESSION['sud']['userID'].";";
-	$dData = db_query($query)[0];
+	foreach ($bmID as $key => $value) {
+		$bm = $value;
+		$query = "SELECT `bmParentID`, `bmIndex`, `bmURL` FROM `bookmarks` WHERE `bmID` = '$bm' AND `userID` = ".$_SESSION['sud']['userID'].";";
+		$dData = db_query($query)[0];
 
-	$query = "DELETE FROM `bookmarks` WHERE `bmID` = '$bmID' AND `userID` = ".$_SESSION['sud']['userID'].";";
-	$count = db_query($query);
+		if(isset($dData['bmType']) && $dData['bmType'] === 'folder') {
+			e_log(8,"'$bmID' appeared to be a folder, delete all the contents of it");
+			$query = "DELETE FROM `bookmarks` WHERE `bmParentID` = '$bm' AND `userID` = ".$_SESSION['sud']['userID'].";";
+			db_query($query);
+		}
 
-	if(isset($dData['bmType']) && $dData['bmType'] === 'folder') {
-		e_log(8,"'$bmID' appeared to be a folder, delete all the contents of it");
-		$query = "DELETE FROM `bookmarks` WHERE `bmParentID` = '$bmID' AND `userID` = ".$_SESSION['sud']['userID'].";";
-		db_query($query);
-	}
+		$query = "DELETE FROM `bookmarks` WHERE `bmID` = '$bm' AND `userID` = ".$_SESSION['sud']['userID'].";";
+		$count = db_query($query);
 
-	e_log(8,"Check for remaining entries in this folder");
-	$query = "SELECT * FROM `bookmarks` WHERE `bmParentID` = '".$dData['bmParentID']."' AND `userID` = ".$_SESSION['sud']['userID']." AND `bmIndex` > ".$dData['bmIndex']." ORDER BY bmIndex;";
-	$fBookmarks = db_query($query);
-	$bm_count = count($fBookmarks);
-	
-	e_log(8,"Re-index folder ".$dData['bmParentID']);
-	for ($i = 0; $i < $bm_count; $i++) {
-		$query = "UPDATE `bookmarks` SET `bmIndex`= $i WHERE `bmID` = '".$fBookmarks[$i]['bmID']."' AND `userID` = ".$_SESSION['sud']['userID'].";";
-		db_query($query);
+		e_log(8,"Check for remaining entries in folder");
+		$query = "SELECT * FROM `bookmarks` WHERE `bmParentID` = '".$dData['bmParentID']."' AND `userID` = ".$_SESSION['sud']['userID']." ORDER BY bmIndex;";
+		$fBookmarks = db_query($query);
+
+		$bm_count = count($fBookmarks);
+		e_log(8, "Re-index folder ".$dData['bmParentID']);
+		for ($i = 0; $i < $bm_count; $i++) {
+			$data[] = array($i, $fBookmarks[$i]['bmID']);
+		}
+
+		$query = "UPDATE `bookmarks` SET `bmIndex` = ? WHERE bmID = ?";
+		db_query($query, $data);
 	}
 
 	return $count;
@@ -1896,46 +1189,6 @@ function addBookmark($bm) {
 	}
 }
 
-function getChanges($cl, $ct, $time) {
-	$uid = $_SESSION['sud']["userID"];
-	e_log(8,"Browser startup sync started, get client data");
-	$query = "SELECT `lastseen` FROM `clients` WHERE `cid` = '".$cl."' AND `uid` = $uid AND `ctype` = '".$ct."';";
-	$clientData = db_query($query)[0];
-
-	if($clientData) {
-		$lastseen = $clientData["lastseen"];
-		e_log(8,"Get changed bookmarks for client $cl");
-		$query = "SELECT a.`bmParentID` as fdID, (SELECT `bmTitle` FROM `bookmarks` WHERE `bmID` = a.`bmParentID` AND userID = $uid) as fdName, (SELECT `bmIndex` FROM `bookmarks` WHERE `bmID` = a.`bmParentID` AND userID = $uid) as fdIndex, `bmID`, `bmIndex`, `bmTitle`, `bmType`, `bmURL`, `bmAdded`, `bmModified` FROM `bookmarks` a WHERE (bmAdded > $lastseen AND userID = $uid) OR (bmAdded > $lastseen AND userID = $uid);";
-		
-		$bookmarkData = db_query($query);
-		foreach($bookmarkData as $key => $entry) {
-			$bookmarkData[$key]['bmTitle'] = html_entity_decode($entry['bmTitle'], ENT_QUOTES, 'UTF-8'); 
-		}
-	}
-	else {
-		e_log(2,"Client not found in database, registering now");
-		updateClient($cl, $ct, $time, true);
-		return "New client registered for user.";
-	}
-
-	if (!empty($bookmarkData)) {
-		updateClient($cl, $ct, $time, true);
-
-		if(CONFIG['cexpjson'] && CONFIG['loglevel'] == 9) {
-			$filename = is_dir(CONFIG['logfile']) ? CONFIG['logfile']."/changes_".time().".json":"changes_".time().".json";
-			file_put_contents($filename,json_encode($bookmarkData),true);
-			e_log(8,'Export file is saved to '.$filename);
-		}
-
-		e_log(8,"Found ".count($bookmarkData)." changes. Sending them to the client");
-		return $bookmarkData;
-	}
-	else {
-		e_log(8,"No bookmarks changed since last sync");
-		return "No bookmarks added, removed or changed since the client was last seen.";
-	}
-}
-
 function updateClient($cl, $ct, $time, $sync = false) {
 	$fclients = array("bookmarkTab", "Android");
 	if(in_array($cl, $fclients)) return 0;
@@ -1976,7 +1229,7 @@ function bmTree() {
 
 function getIndex($folder) {
 	e_log(8,"Get new bookmark ID");
-	$query = "SELECT MAX(`bmIndex`) AS OIndex  FROM `bookmarks` WHERE `bmParentID` = '".$folder."'";
+	$query = "SELECT IFNULL(MAX(`bmIndex`), 0) AS OIndex  FROM `bookmarks` WHERE `bmParentID` = '".$folder."'";
 	$IndexArr = db_query($query);
 	$maxIndex = $IndexArr[0]['OIndex'] + 1;
 	return $maxIndex;
@@ -1986,14 +1239,10 @@ function getSiteTitle($url) {
 	e_log(8,"Get titel for site ".$url);
 	$src = file_get_contents($url);
 	if(strlen($src) > 0) {
-		preg_match("/\<title\>(.*)\<\/title\>/i",$src,$title_arr);
-		$title = (strlen($title_arr[1]) > 0) ? strval($title_arr[1]) : substr($url, 0, 240);
-		e_log(8,"Titel for site is '$title'");
-		$convTitle = htmlspecialchars(mb_convert_encoding(htmlspecialchars_decode($title, ENT_QUOTES),"UTF-8"),ENT_QUOTES,'UTF-8', false);
-	} else {
-		$convTitle = substr($url, 0, 240);
+		$title = preg_match('/<title[^>]*>(.*?)<\/title>/ims', $src, $matches) ? $matches[1] : null;
+		$convTitle = ($title == '' ) ? substr($url, 0, 240):htmlspecialchars(mb_convert_encoding(htmlspecialchars_decode($title, ENT_QUOTES),"UTF-8"),ENT_QUOTES,'UTF-8', false);
 	}
-	
+	e_log(8,"Titel for site is '$convTitle'");
 	return $convTitle;
 }
 
@@ -2055,16 +1304,6 @@ function e_log($level, $message, $errfile="", $errline="", $output=0) {
 	}
 }
 
-function filterIP($remote) {
-	$v4mapped_prefix_bin = hex2bin('00000000000000000000ffff'); 
-	$addr_bin = inet_pton($remote);
-	if($addr_bin === FALSE ) die(e_log(1,'Invalid IP address'));
-
-	if( substr($addr_bin, 0, strlen($v4mapped_prefix_bin)) == $v4mapped_prefix_bin) $addr_bin = substr($addr_bin, strlen($v4mapped_prefix_bin));
-
-	return inet_ntop($addr_bin);
-}
-
 function delUsermarks($uid) {
 	e_log(8, "Delete all bookmarks for logged in user");
 	$query = "DELETE FROM `bookmarks` WHERE `UserID`=".$uid;
@@ -2076,7 +1315,7 @@ function htmlHeader() {
 		<html lang='en'>
 			<head>
 				<meta name='viewport' content='width=device-width, initial-scale=1'>
-				<script src='js/bookmarks.js'></script>
+				<script src='js/bookmarks.min.js'></script>
 				<link type='text/css' rel='stylesheet' href='css/bookmarks.min.css'>
 				<link rel='shortcut icon' type='image/x-icon' href='images/bookmarks.ico'>
 				<link rel='manifest' href='manifest.json'>
@@ -2267,7 +1506,7 @@ function htmlForms() {
 	$footerButton = "
 	<div id='bmarkadd' class='mbmdialog'>
 		<h6>Add Bookmark</h6>
-		<form id='bmadd' action='?madd' method='POST'>
+		<form id='bmadd' action='?' method='POST'>
 			<input placeholder='URL' type='text' id='url' name='url' value=''>
 			<div class='select'>
 				<select id='folder' name='folder'>
@@ -2275,7 +1514,7 @@ function htmlForms() {
 				</select>
 				<div class='select__arrow'></div>
 			</div>
-			<div class='dbutton'><button type='submit' id='save' name='madd' value='Save'>Save</button></div>
+			<div class='dbutton'><button type='submit' id='save' name='' value='Save'>Save</button></div>
 		</form>
 	</div>
 	<div id='footer'></div>";
@@ -2332,7 +1571,7 @@ function notiList($uid, $loop) {
 }
 
 function htmlFooter() {
-	$htmlFooter = "<script src='js/bookmarksf.js'></script></body></html>";
+	$htmlFooter = "<script src='js/bookmarksf.min.js'></script></body></html>";
 	return $htmlFooter;
 }
 
@@ -2915,13 +2154,13 @@ function checkDB() {
 		e_log(8,"Database update needed. Starting DB update...");
 		if(CONFIG['db']['type'] == "sqlite") {
 			db_query(file_get_contents("./sql/sqlite_update_$dbv.sql"));
-		} elseif(CONFIG['db']['type'] == "mysql") {
+		} elseif (CONFIG['db']['type'] == "mysql") {
 			db_query(file_get_contents("./sql/mysql_update_$dbv.sql"));
 		}
 		$aversion = explode ("\n", file_get_contents('./CHANGELOG.md',NULL,NULL,0,30))[2];
 	    $aversion = substr($aversion,0,strpos($aversion, " "));
 		db_query("INSERT INTO `system`(`app_version`,`db_version`,`updated`) VALUES ('$aversion','$dbv','$newdate');");
-	} elseif($vInfo['db_version'] && $vInfo['db_version'] >= $dbv) {
+	} elseif ($vInfo['db_version'] && $vInfo['db_version'] >= $dbv) {
 		if($olddate <> $newdate) db_query("UPDATE `system` SET `updated` = '$newdate' WHERE `updated` = '$olddate';");
 	} else {
 		e_log(2,"Database not ready. Initialize database now");
@@ -2941,7 +2180,7 @@ function checkDB() {
 				e_log(8,"Initialise new SQLite database");
 				db_query(file_get_contents("./sql/sqlite_init.sql"));
 			}
-		} elseif(CONFIG['db']['type'] == "mysql") {
+		} elseif (CONFIG['db']['type'] == "mysql") {
 			e_log(8,"Initialise new MySQL database");
 			db_query(file_get_contents("./sql/mysql_init.sql"));
 		}