+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ChangeLog
 =========
+1.6.8 (2022-05-04)
+-------------------------
+- Reworked actions
+- Reworked XHR requests
+- Removed unused functions
+- Added delete multiple bookmarks (Ctrl + Click > Del)
+- Fix Re-Index folder
+- Fix Rename client
+- Other Small Fixes
+
+
 1.6.7 (2022-03-19)
 -------------------------
 - Fixed delete bookmarks